Adds a vue to manage tickets withdrawal
This adds a vue where you can select a user who has tickets,, a type of ticket and remove a select nuimber of  tickets from them. It also updates the last time a user came to the coworking space.
Also adds a few comments to the trigger file and a very short description of the module's features inside the readme file.